Move methods in LadderFormat class to GameUtils and delete LadderFormat
